Moved pagination.py out of models dir, renamed some variables for more clarity
